+++ test/ecto/changeset/belongs_to_test.exs
@@ -192,14 +192,14 @@ defmodule Ecto.Changeset.BelongsToTest do
     end
   end
 
-  test "cast belongs_to with :invalid parameters" do
-    changeset = cast(%Author{profile: nil}, :invalid, :profile)
+  test "cast belongs_to with empty parameters" do
+    changeset = cast(%Author{profile: nil}, %{}, :profile)
     assert changeset.changes == %{}
 
-    changeset = cast(%Author{}, :invalid, :profile, required: true)
+    changeset = cast(%Author{}, %{}, :profile, required: true)
     assert changeset.changes == %{}
 
-    changeset = cast(%Author{profile: %Profile{}}, :invalid, :profile, required: true)
+    changeset = cast(%Author{profile: %Profile{}}, %{}, :profile, required: true)
     assert changeset.changes == %{}
   end